+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   "decorator": "3.18.0",
   "core": "3.18.0",
-  "timestamp": 1727087291497
+  "timestamp": 1728826525934
 };