+++ scripts/rollback/rollback-3.18.2.sh
@@ -1,0 +1,124 @@
+#!/bin/bash
+# timestamp: 1728826359639
+
+npm dist-tag add @midwayjs/fc-starter@3.18.0 latest
+npm dist-tag add @midwayjs/axios@3.18.0 latest
+npm dist-tag add @midwayjs/bootstrap@3.18.0 latest
+npm dist-tag add @midwayjs/bull@3.18.0 latest
+npm dist-tag add @midwayjs/bull-board@3.18.0 latest
+npm dist-tag add @midwayjs/busboy@3.18.0 latest
+npm dist-tag add @midwayjs/cache-manager@3.18.1 latest
+npm dist-tag add @midwayjs/captcha@3.18.1 latest
+npm dist-tag add @midwayjs/casbin@3.18.0 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.18.0 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.18.0 latest
+npm dist-tag add @midwayjs/code-dye@3.18.0 latest
+npm dist-tag add @midwayjs/consul@3.18.0 latest
+npm dist-tag add @midwayjs/cos@3.18.0 latest
+npm dist-tag add @midwayjs/cron@3.18.0 latest
+npm dist-tag add @midwayjs/cross-domain@3.18.0 latest
+npm dist-tag add @midwayjs/etcd@3.18.0 latest
+npm dist-tag add @midwayjs/express-session@3.18.0 latest
+npm dist-tag add @midwayjs/faas@3.18.0 latest
+npm dist-tag add @midwayjs/grpc@3.18.0 latest
+npm dist-tag add @midwayjs/http-proxy@3.18.0 latest
+npm dist-tag add @midwayjs/i18n@3.18.0 latest
+npm dist-tag add @midwayjs/info@3.18.0 latest
+npm dist-tag add @midwayjs/jwt@3.18.0 latest
+npm dist-tag add @midwayjs/kafka@3.18.0 latest
+npm dist-tag add @midwayjs/leoric@3.18.0 latest
+npm dist-tag add @midwayjs/mikro@3.18.0 latest
+npm dist-tag add @midwayjs/mock@3.18.0 latest
+npm dist-tag add @midwayjs/mongoose@3.18.0 latest
+npm dist-tag add @midwayjs/mqtt@3.18.0 latest
+npm dist-tag add @midwayjs/oss@3.18.0 latest
+npm dist-tag add @midwayjs/otel@3.18.0 latest
+npm dist-tag add @midwayjs/passport@3.18.0 latest
+npm dist-tag add @midwayjs/process-agent@3.18.0 latest
+npm dist-tag add @midwayjs/prometheus@3.18.0 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.18.0 latest
+npm dist-tag add @midwayjs/rabbitmq@3.18.0 latest
+npm dist-tag add @midwayjs/redis@3.18.0 latest
+npm dist-tag add @midwayjs/security@3.18.0 latest
+npm dist-tag add @midwayjs/sequelize@3.18.0 latest
+npm dist-tag add @midwayjs/session@3.18.0 latest
+npm dist-tag add @midwayjs/socketio@3.18.0 latest
+npm dist-tag add @midwayjs/static-file@3.18.0 latest
+npm dist-tag add @midwayjs/swagger@3.18.1 latest
+npm dist-tag add @midwayjs/tablestore@3.18.0 latest
+npm dist-tag add @midwayjs/tags@3.18.0 latest
+npm dist-tag add @midwayjs/tenant@3.18.0 latest
+npm dist-tag add @midwayjs/typegoose@3.18.0 latest
+npm dist-tag add @midwayjs/typeorm@3.18.0 latest
+npm dist-tag add @midwayjs/upload@3.18.0 latest
+npm dist-tag add @midwayjs/validate@3.18.0 latest
+npm dist-tag add @midwayjs/version@3.18.1 latest
+npm dist-tag add @midwayjs/view@3.18.0 latest
+npm dist-tag add @midwayjs/view-ejs@3.18.0 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.18.0 latest
+npm dist-tag add @midwayjs/web@3.18.0 latest
+npm dist-tag add @midwayjs/express@3.18.0 latest
+npm dist-tag add @midwayjs/koa@3.18.0 latest
+npm dist-tag add @midwayjs/ws@3.18.0 latest
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.18.0 => 3.18.2
+#  - @midwayjs/axios: 3.18.0 => 3.18.2
+#  - @midwayjs/bootstrap: 3.18.0 => 3.18.2
+#  - @midwayjs/bull: 3.18.0 => 3.18.2
+#  - @midwayjs/bull-board: 3.18.0 => 3.18.2
+#  - @midwayjs/busboy: 3.18.0 => 3.18.2
+#  - @midwayjs/cache-manager: 3.18.1 => 3.18.2
+#  - @midwayjs/captcha: 3.18.1 => 3.18.2
+#  - @midwayjs/casbin: 3.18.0 => 3.18.2
+#  - @midwayjs/casbin-redis-adapter: 3.18.0 => 3.18.2
+#  - @midwayjs/casbin-typeorm-adapter: 3.18.0 => 3.18.2
+#  - @midwayjs/code-dye: 3.18.0 => 3.18.2
+#  - @midwayjs/consul: 3.18.0 => 3.18.2
+#  - @midwayjs/cos: 3.18.0 => 3.18.2
+#  - @midwayjs/cron: 3.18.0 => 3.18.2
+#  - @midwayjs/cross-domain: 3.18.0 => 3.18.2
+#  - @midwayjs/etcd: 3.18.0 => 3.18.2
+#  - @midwayjs/express-session: 3.18.0 => 3.18.2
+#  - @midwayjs/faas: 3.18.0 => 3.18.2
+#  - @midwayjs/grpc: 3.18.0 => 3.18.2
+#  - @midwayjs/http-proxy: 3.18.0 => 3.18.2
+#  - @midwayjs/i18n: 3.18.0 => 3.18.2
+#  - @midwayjs/info: 3.18.0 => 3.18.2
+#  - @midwayjs/jwt: 3.18.0 => 3.18.2
+#  - @midwayjs/kafka: 3.18.0 => 3.18.2
+#  - @midwayjs/leoric: 3.18.0 => 3.18.2
+#  - @midwayjs/mikro: 3.18.0 => 3.18.2
+#  - @midwayjs/mock: 3.18.0 => 3.18.2
+#  - @midwayjs/mongoose: 3.18.0 => 3.18.2
+#  - @midwayjs/mqtt: 3.18.0 => 3.18.2
+#  - @midwayjs/oss: 3.18.0 => 3.18.2
+#  - @midwayjs/otel: 3.18.0 => 3.18.2
+#  - @midwayjs/passport: 3.18.0 => 3.18.2
+#  - @midwayjs/process-agent: 3.18.0 => 3.18.2
+#  - @midwayjs/prometheus: 3.18.0 => 3.18.2
+#  - @midwayjs/prometheus-socket-io: 3.18.0 => 3.18.2
+#  - @midwayjs/rabbitmq: 3.18.0 => 3.18.2
+#  - @midwayjs/redis: 3.18.0 => 3.18.2
+#  - @midwayjs/security: 3.18.0 => 3.18.2
+#  - @midwayjs/sequelize: 3.18.0 => 3.18.2
+#  - @midwayjs/session: 3.18.0 => 3.18.2
+#  - @midwayjs/socketio: 3.18.0 => 3.18.2
+#  - @midwayjs/static-file: 3.18.0 => 3.18.2
+#  - @midwayjs/swagger: 3.18.1 => 3.18.2
+#  - @midwayjs/tablestore: 3.18.0 => 3.18.2
+#  - @midwayjs/tags: 3.18.0 => 3.18.2
+#  - @midwayjs/tenant: 3.18.0 => 3.18.2
+#  - @midwayjs/typegoose: 3.18.0 => 3.18.2
+#  - @midwayjs/typeorm: 3.18.0 => 3.18.2
+#  - @midwayjs/upload: 3.18.0 => 3.18.2
+#  - @midwayjs/validate: 3.18.0 => 3.18.2
+#  - @midwayjs/version: 3.18.1 => 3.18.2
+#  - @midwayjs/view: 3.18.0 => 3.18.2
+#  - @midwayjs/view-ejs: 3.18.0 => 3.18.2
+#  - @midwayjs/view-nunjucks: 3.18.0 => 3.18.2
+#  - @midwayjs/web: 3.18.0 => 3.18.2
+#  - @midwayjs/express: 3.18.0 => 3.18.2
+#  - @midwayjs/koa: 3.18.0 => 3.18.2
+#  - @midwayjs/ws: 3.18.0 => 3.18.2